requirements.txt
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
ultralytics>=8.0.0
